move factory response logic from JavaNetSinkHttpClient to HttpResponseChecker
